+++ integration-tests/cypress/e2e/using-ya-to-read-events/reading-events.cy.ts
@@ -112,14 +112,14 @@ describe("'rename' events", () => {
       cy.get("Rename").should("not.exist")
 
       // yazi should be showing the new file name
-      const file = dir.contents["initial-file.txt"]
-      cy.contains(`${file.stem}2${file.extension}`)
+      const newFileName = `initial-file2.txt`
+      cy.contains(newFileName)
 
       // close yazi
       cy.typeIntoTerminal("q")
 
       // the buffer name should now be updated
-      cy.contains(`${file.stem}2${file.extension}`)
+      cy.contains(newFileName)
     })
   })
 
@@ -140,26 +140,58 @@ describe("'rename' events", () => {
       cy.get("Rename").should("not.exist")
 
       // yazi should be showing the new file name
-      const file = dir.contents["initial-file.txt"]
-      cy.contains(`${file.stem}2${file.extension}`)
+      const newFileName = "initial-file2.txt"
+      cy.contains(newFileName)
 
       // close yazi
       cy.typeIntoTerminal("q")
 
-      const newName = `${file.stem}2${file.extension}`
       // the buffer name should now be updated
-      cy.contains(newName)
+      cy.contains(newFileName)
 
       // rename a second time, returning to the original name
       cy.typeIntoTerminal("{upArrow}")
       cy.typeIntoTerminal("r")
       cy.contains("Rename:")
       cy.typeIntoTerminal("{backspace}")
-      cy.contains(`${file.stem}${file.extension}`)
+      cy.contains(newFileName)
       cy.typeIntoTerminal("{enter}")
 
       cy.typeIntoTerminal("q")
-      cy.contains(newName).should("not.exist")
+      cy.contains(newFileName).should("not.exist")
+    })
+  })
+
+  it("can publish YaziRenamedOrMoved events when a file is renamed", () => {
+    cy.startNeovim({
+      startupScriptModifications: ["notify_rename_events.lua"],
+    }).then((dir) => {
+      // the default file should already be open
+      cy.contains(dir.contents["initial-file.txt"].name)
+      cy.contains("If you see this text, Neovim is ready!")
+
+      // start yazi
+      cy.typeIntoTerminal("{upArrow}")
+
+      // start file renaming
+      cy.typeIntoTerminal("r")
+      cy.contains("Rename:")
+      cy.typeIntoTerminal("2{enter}")
+
+      cy.get("Rename").should("not.exist")
+
+      // yazi should be showing the new file name
+      const newFileName = `initial-file2.txt`
+      cy.contains(newFileName)
+
+      // close yazi
+      cy.typeIntoTerminal("q")
+
+      // yazi should now be closed
+      cy.contains("-- TERMINAL --").should("not.exist")
+
+      cy.typeIntoTerminal(":mes{enter}")
+      cy.contains("Just received a YaziRenamedOrMoved event!")
     })
   })
 })