+++ lua/yazi/keybinding_helpers.lua
@@ -212,4 +212,25 @@ function YaziOpenerActions.replace_in_selected_files(config, chosen_files)
   end
 end
 
+---@param context YaziActiveContext
+function YaziOpenerActions.change_working_directory(context)
+  local last_directory = context.ya_process.cwd
+  if not last_directory then
+    assert(
+      context.input_path,
+      "No input_path found. Expected yazi to be started with an input_path"
+    )
+    if context.input_path:is_file() then
+      last_directory = context.input_path:parent().filename
+    else
+      last_directory = context.input_path.filename
+    end
+  end
+
+  if last_directory ~= vim.fn.getcwd() then
+    vim.notify('cwd changed to "' .. last_directory .. '"')
+    vim.cmd({ cmd = "cd", args = { last_directory } })
+  end
+end
+
 return YaziOpenerActions